refactor(cli): flatten phone command module
Move the Phone command slice to the neutral command root and update its internal export without changing public CLI or AgentPhone behavior.
